delete old pages + move Cards
