--- app/code/data.ts
@@ -1,115 +1,0 @@
-export const codeData = [
-  {
-    title: "Fake Stack Overflow",
-    thumbnail: "/assets/code/fake-stack-overflow-mockup.png",
-    links: [
-      {
-        type: "live",
-        url: "https://cs4530-s25-404.onrender.com/",
-      },
-    ],
-    description:
-      "A full-stack Stack Overflow clone with Single Sign-On (SSO) authentication and natural language searching features. UI refreshed using Chakra UI.",
-    technologies: [
-      "React",
-      "TypeScript",
-      "Express.js",
-      "MongoDB",
-      "Jest",
-      "Chakra UI",
-    ],
-  },
-  {
-    title: "Vet Finder",
-    links: [
-      {
-        type: "github",
-        url: "https://github.com/chanmioh/team-animal",
-      },
-    ],
-    thumbnail: "/assets/code/chewy-mockup.png",
-    description:
-      "A user-friendly vet finder with filters. Built for and awarded 1st place at Chewy's 2023 Diversity Hackathon.",
-    technologies: ["Next.js", "JavaScript", "Tailwind CSS"],
-  },
-  {
-    title: "FinishLine",
-    links: [
-      {
-        type: "github",
-        url: "https://github.com/Northeastern-Electric-Racing/FinishLine",
-      },
-      {
-        type: "live",
-        url: "https://finishlinebyner.com",
-      },
-    ],
-    thumbnail: "/assets/code/finishline-mockup.png",
-    description:
-      "Project management dashboard for Northeastern Electric Racing.",
-    technologies: ["React", "TypeScript", "Express.js", "Prisma", "Jest"],
-  },
-  {
-    title: "Image Processor",
-    links: [
-      {
-        type: "github",
-        url: "https://github.com/sharonyang16/image-processing",
-      },
-    ],
-    thumbnail: "/assets/code/image-mockup.png",
-    description: "Image processing and manipulation app with various filters.",
-    technologies: ["Java", "JUnit", "Java Swing"],
-  },
-  {
-    title: "Typing Game",
-    links: [
-      {
-        type: "github",
-        url: "https://github.com/sharonyang16/typing-game",
-      },
-      {
-        type: "live",
-        url: "https://typing-game-navy-nine.vercel.app/",
-      },
-    ],
-    thumbnail: "/assets/code/typing-mockup.png",
-    description: "A full stack typing game with test scores and leaderboards.",
-    technologies: [
-      "Next.js",
-      "TypeScript",
-      "Express.js",
-      "Prisma",
-      "Tailwind CSS",
-      "Firebase",
-    ],
-  },
-  {
-    title: "Rosy Photobooth",
-    links: [
-      {
-        type: "github",
-        url: "https://github.com/sharonyang16/rosy-photo-booth",
-      },
-      {
-        type: "live",
-        url: "https://rosy-photo-booth.vercel.app/",
-      },
-    ],
-    thumbnail: "/assets/code/rosy-mockup.png",
-    description: "A Korean-style Photo Booth app.",
-    technologies: ["Next.js", "TypeScript", "Tailwind CSS"],
-  },
-  {
-    title: "Timbers' Nap",
-    links: [
-      {
-        type: "live",
-        url: "https://play.unity.com/mg/other/webgl-builds-349008",
-      },
-    ],
-    thumbnail: "/assets/code/timbers-mockup.png",
-    description: "3D-platformer with 3 engaging levels. ",
-    technologies: ["Unity3D", "C#"],
-  },
-];
