--- app/design/data.ts
@@ -1,64 +1,0 @@
-export const designData = [
-  {
-    title: "Lift Off",
-    links: [
-      {
-        type: "figma",
-        url: "https://www.figma.com/proto/wzqtjLJcs8259BZL5lLUfv/Lift-Off?page-id=446%3A3941&viewport=322%2C371%2C0.2&t=dDE1iOnE0Aa7YqsH-1&scaling=scale-down&content-scaling=fixed&node-id=446-4778&starting-point-node-id=446%3A4778",
-      },
-    ],
-    thumbnail: "/assets/design/lift-off-mockups.png",
-    description: "Strength training log book app.",
-    technologies: ["Mobile", "Figma", "Illustrator", "User Research"],
-  },
-  {
-    title: "Oven Buddy",
-    links: [
-      {
-        type: "figma",
-        url: "https://www.figma.com/proto/F6k50RNQ4dFm1bBN2cDdTQ/T---Sharon-Yang?page-id=1%3A5&node-id=6110-2517&viewport=380%2C-1696%2C0.69&t=IdlP3yqxew7CWyuE-1&scaling=scale-down&content-scaling=fixed&starting-point-node-id=6110%3A2517",
-      },
-    ],
-    thumbnail: "/assets/design/oven-buddy-mockups.png",
-    description: "Uber inspired smart oven app.",
-    technologies: ["Mobile", "Figma", "User Research"],
-  },
-  {
-    title: "Apple Maps Explore",
-    links: [
-      {
-        type: "figma",
-        url: "https://www.figma.com/proto/QRPEdURbvR8RSvTiDIK2kF/ARTG2400---Final-Project?page-id=2%3A2&node-id=27-1081&viewport=434%2C-274%2C0.39&t=hPdyOe6DSO0ZyehG-1&scaling=scale-down&content-scaling=fixed&starting-point-node-id=27%3A1081",
-      },
-    ],
-    thumbnail: "/assets/design/apple-maps-adventure-mockups.png",
-    description:
-      "A new apple maps feature focused on exploring new places and making navigation an adventure.",
-    technologies: ["Mobile", "Figma", "User Research"],
-  },
-  {
-    title: "Argos Dashboard Revamp",
-    thumbnail: "/assets/design/argos-mockups.png",
-    description:
-      "Redesigned the home page of the car data visualization dashboard to have a more interesting and modular design.",
-    technologies: ["Desktop", "Mobile", "Figma", "User Research"],
-  },
-  {
-    title: "NUSkillz",
-    description:
-      "Social media app for developing new skills, focusing on connecting users with like-minded peers.",
-    technologies: ["Desktop", "Figma", "User Research"],
-  },
-  {
-    title: "FinishLine Guest Mode/Mobile Redesign",
-    description:
-      "Focused on transforming existing project managmenet dashboard UI to be more useful for passing viewers and more responsive on smaller screens.",
-    technologies: ["Mobile", "Figma", "User Research"],
-  },
-  {
-    title: "FinishLine Home Page Revamp",
-    description:
-      "Redesigned the home page of the internal project management dashboard to be more useful for members of different roles.",
-    technologies: ["Desktop", "Figma"],
-  },
-];
